refactor/#293 프로토콜 및 서비스타입 리팩
